Fernando Pessoa University 3D Interactive Building
This web application consists of allowing users (i.e., professors and students) to check in real-time the
rooms occupancies navigating through a 3D interactive building of the university, and access/review
the characteristics of each rooms. This way, given a specific date and hour, they could reserve the
rooms they want, if they were free. The admin would be able to connect to the schedule database,
directly from the purpose web application, and edit all of this information.